src/main/java/com/greencity/service/impl/ProductCategoryServiceImpl.java
package com.greencity.service.impl;

import com.greencity.dao.ProductCategoryDao;
import com.greencity.entity.ProductCategory;
import com.greencity.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

/**
 * 类目
 * author： Andyfung
 * time： 2020.07.02
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryDao productCategoryDao;

    @Override
    public Optional<ProductCategory> findOne(Integer gategoryId) {
        return  productCategoryDao.findById(gategoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryDao.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
